Revert "Introduzindo axios"
