--- front-end/src/services/index.js
@@ -1,1 +1,0 @@
-export { companyService as default } from './company';
